feat: add new feature

+++ sipelatihan/src/main/java/apap/ta/sipelatihan/Repository/PelatihanDb.java
@@ -7,6 +7,6 @@ import java.util.List;
 import java.util.Optional;
 
 
-public interface PelatihanDb extends JpaRepository<PelatihanModel, Long>{
+public interface PelatihanDb extends JpaRepository<PelatihanModel, Integer>{
     List<PelatihanModel> findAllByOrderById();
 }
