+++ sipelatihan/src/main/java/apap/ta/sipelatihan/repository/TrainerDb.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 
-public interface TrainerDb extends JpaRepository<TrainerModel, Long>{
+public interface TrainerDb extends JpaRepository<TrainerModel, Integer>{
     TrainerModel findTrainerModelById(Integer id);
 
     TrainerModel findByNoKtp(String noKtp);
