+++ sipelatihan/src/main/java/apap/ta/sipelatihan/repository/UserDb.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserDb extends JpaRepository<UserModel,Long> {
+public interface UserDb extends JpaRepository<UserModel, Integer> {
     UserModel findByUsername(String username);
 }
